+++ src/components/techology/techology_items.tsx
